Criando um menur de opcao r

+++ ex059.py
@@ -7,8 +7,8 @@
 
 menu = 4
 while menu != 5:
-    n1 = float(input('Digite o 1º número: '))
-    n2 = float(input('Digite o 2º número: '))
+    n1 = int(input('Digite o 1º número: '))
+    n2 = int(input('Digite o 2º número: '))
     menu = int(input('''
     ---------- Opção ----------
     [1]somar
